feat: Upgrade to optimized jnigen 3.1-SNAPSHOT

+++ gdx-box2d-utils/src/main/java/com/badlogic/gdx/box2d/utils/Box2dWorldTaskSystem.java
@@ -103,7 +103,7 @@ public class Box2dWorldTaskSystem implements Disposable {
                     start = end;
                 }
 
-                VoidPointer taskIndex = new VoidPointer(taskCount + 1, false);
+                VoidPointer taskIndex = new VoidPointer((long) taskCount + 1, false);
                 taskCount++;
                 return taskIndex;
             } else {
